\name{spatialSEIR-package}
\alias{spatialSEIR-package}
\alias{spatialSEIR}
\docType{package}
\title{
A fast framework for fitting Bayesian spatial SEIR models. 
~~ spatialSEIR ~~
}
\description{
Compartmental epidemic models are flexible and intuitive tools for
modeling the temporal evolution of epidemics, and can be easily extended to 
the spatio-temporal world. On the other hand, the computational burden grows 
greatly for large, or highly interconnected, spatial domains. This package 
aims to provide a fast method of fitting such models by shifting as much 
of the computational burden as possible to graphics processors using OpenCL. 
}
\details{
\tabular{ll}{
Package: \tab spatialSEIR\cr
Type: \tab Package\cr
Version: \tab 0.01\cr
Date: \tab 2014-02-26\cr
License: \tab GPL3\cr
}


}
\author{
Grant Brown

Maintainer: Grant Brown <grant.brown73@gmail.com>
}
\references{
Coming Soon
}


\keyword{ package }
\seealso{
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
    # Examples to come
}
